io: Make peer_addr fallible
The io::PeerAddr trait assumes that the peer addr lookup is infallible,
panicking the process when the call fails. In practice, however, this
call can fail when a system is under load.

This change modifies the `io::PeerAddr` trait to allow this lookup to
fail. When it fails, we propagate this failure to the connection's task.