metrics: Expose process_cpu_seconds_total as a float (#754)
Prometheus handles all values as `f64`, but we only expose values as
whole integers. This means that the `process_cpu_seconds_total` metric
only exposes whole second values, while Linux exposes process time in
10ms increments.

This change modifies the `Counter` metric type to store an additional
marker that provides a strategy for converting the stored `u64` value to
`f64` for export.  This strategy is employed so that we can continue to
use `AtomicU64` to back counters and only use floats at export-time.  By
default the unit type is used to convert counters as before, but an
alternate `MillisAsSeconds` strategy is used to expose fractional
seconds from a millisecond counter.

This necessitates changing the histogram buckets to floats as well.
While this change doesn't modify the bucket values, this sets up future
changes to latency metrics.