Audit uses of SmallRng
While testing another program, I was reminded that `SmallRng` can have
surprising behavior when cloned: since a SmallRng's internal state is
stored in a field, each clone ends up in the same state such that each
clone produces the same sequence of values.

Furthermore, the use of `SmallRng::from_entropy` uses `getrandom`
internally, which can be quite slow, especially compared to the
alternative `SmallRng::from_rng(&mut thread_rng())`.

It doesn't appear that any of our uses of `SmallRng` are particularly
problematic -- i.e. the services that hold a `SmallRng` are not cloned,
so we shouldn't be in a place where we always use a single value.
However, in the spirit of being defensive to this scenario, this change
avoid storing a `SmallRng` in Layer and NewService implemetnations so
that each produced service obtains a unique `SmallRng`. Furthermore, all
uses of `SmallRng::from_entropy()` are converted to use the faster
thread-local PRNG equivalent.