cache: Decouple tracker from cache layer
The cache layer takes another layer that is responsible for dropping
each cache entry's handle so that it can be evicted; but now that
there's a single connection-oriented cache on the outbound stack, cache
retention is tied to the time since the last connection and not the
activity of the conncetion.

This change doesn't fix this underlying issue, but it seperates the
tracker layer from the cache layer so that a followup change can modify
the cache retention strategy.