cache: Move idle eviction into the tracking layer
The cache relies on the buffer's idle timeout to evict cache entries.
This approach (1) causes us to rely on a specialized Buffer
implementation (rather than using tower's Buffer) and (2) is inflexible,
preventing the outbound connection-oriented cache from `taking into
account connection-level activity.

This change modifies the `cache::NewTrack` middleware to spawn a task
that tracks its service's idleness. When the service hasn't been used
for an idle timeout, the cache handle is dropped so that the cache knows
to clear the appropriate entry.