Ensure that services are held as long they are being used
Middlewares, especially the cache, may want to use RAII to detect when a
service is idle, but the TCP server and HTTP routers drop services as
soon as the request is dispatched.

This change modifies the HTTP server to hold the TCP stack until all
work on that connection is complete. It also introduces a new
`http::Retain` middleware that clones its inner service into repsonse
bodies.

This is necesarry for upcoming cache eviction changes to address
linkerd/linkerd2#5334.