outbound: add stack tests for http (#765)
This branch adds some new stack tests for simple HTTP/1 and HTTP/2
meshed and unmeshed outbound requests. These tests also assert that TCP
load balancers are not constructed when HTTP traffic is detected.

The hope was that the panic on constructing TCP balancers would catch a
bug that exists in #744, but that issue appears to be related to the
cache instead. Still, it's probably worth having these tests.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>