cache: Only spawn a single task per cache entry
In b0cf2fbb, the proxy moved eviction to be controlled by a background
task; but this scheme spawned a task every time a service was dropped.

This change modifies the cache to only spawn a single task per cached
service and use a notification channel to signal drops. This reduces
memory pressure, especially for HTTP caches.