detect: Make protocol detection more robust
The existing protocol detection implementation is somewhat naive: we
simply perform a single read on the socket and use that initial buffer
to try to determine the protocol. This means that if the initial read
only returns a few bytes, we won't actually have enough information to
do protocol detection. Furthermore, because we don't actually read until
a `\r\n`, we can incorrectly infer that a non-HTTP protocol is HTTP/1.1.

This change fixes this by replacing the `Prefix` middleware with a
`Detect` service and trait that does its own buffering as it tries to
read from the socket.

When detection times out, an error is thrown to the server. This change
requires modifying test data to include a trailing newline (since it's
not feasible to close these connections given the current test
infrastructure).

This change increases the default dispatch (and therefore detection)
timeouts to 5s.