Introduce a connection-header module
Proxies are not able to securely transport server-first protocols,
since the server-side proxy cannot now whether to perform TLS detection
or direct forwarding. Similarly, multicluster gateways are unable to
proxy arbitrary TCP connections, since the gateway has no means to
discover the intended target for a connection in the destination
cluster.

In order to address both of these situations, this change introduces a
"connection header" module that can be used to transport routing
information for arbitrary TCP streams. When instrumented in the stack,
this module will allow inbound proxies to receive mTLS connections on a
dedicated port and route the connection to proper destination port.
Similar logic applies for multicluster gateways.

This protocol includes a prefix, `"proxy.l5d.io/connect\r\n\r\n"` so
that proxies can differentiate these streams from HTTP streams.
Following this, a network-endian 4-byte length prefix is encoded,
indicating the number of bytes comprising the protocol header. Finally,
the protocol header is encoded as a protobuf structure. The protobuf
message currently includes a target port, which must be set to indicate
how the connection should be routed, as well as an optional `name`,
which may be used for discovery in the multicluster gateway use case.

This change does not actually instrument the module into the proxy
stacks.