trace: update task-tracking to Tokio 0.3 (#778)
This branch updates the `tokio-trace` crate, which we use to implement
the task-tracking endpoint, to a version that depends on Tokio 0.3. This
is necessary for task tracking to work, and also removes an unnecessary
Tokio 0.2 dep from our dependency tree.

Also, the future type name field was removed in Tokio 0.3, since it's
_far_ too long to be useful, and was replaced with spawn locations.
Hopefully, this will make task tracking significantly more useful, and
reduce the memory overhead.