transport: expire idle TCP metrics (#781)
The proxy now exposes transport metrics with per-endpoint metadata; but,
unlike HTTP metrics, these metrics are never dropped. So an application
that has communicated with a now-defunct pod will continue to expose
metrics about that pod indefinitely.

This branch factors out the existing code for expiring idle HTTP metrics
so that it can be used for transport metrics as well, and adds eviction
for transport metrics. 

I also fixed a couple issues in the `linkerd2_metrics::summary` module
that were making the crate fail to compile with `--all-features`.

Closes linkerd/linkerd2#5372.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>