Ensure all forwarded TCP streams keep the proxy running
The proxy is not intended to shutdown until all in-flight responses and
TCP streams are complete. This is enforced by the `ServeHttp`
middleware, which confusingly handles some TCP forwarding in addition to
HTTP server initialization. But there are HTTP streams that are not
instrumented by this middleware and will therefore be abruptly
interrupted when a shutdown signal is received.

This change modifie the `ServeHttp` middleware to only handle HTTP
streams. a `NewOptional` middleware is introduced to conditionally build
a TCP stack when no HTTP type has been detected; and a `drain::Retain`
middleware is added to prevent the drain signal from being released
until TCP streams are completed.