identity: add a metric tracking when the proxy's cert expires
This branch adds a new `certificate_expiration_timestamp_seconds` to the
proxy's Prometheus metric. When identity is enabled, this metric will
report the time at which the proxy's current certificate expires, in
seconds since the Unix epoch.

This is similar to the metric of the same name exported by
`cert-manager`.

Closes linkerd/linkerd2#5379