Default the protocol detection timeout to 10s (#782)
The proxy fails connections that do not complete protocol detection
within a given timeout. This timeout has been derived from the
request-level dispatch timeout; but in practice, many applications may
create idle connections, i.e. in a connection pool, and distribute
requests over them as requests or events come into the application.

In order to better-support these applications, we want to extend the
detection timeout to permit idle connections to be established without
failing them eagerly; though we do not want to increase the
request-level dispatch timeout, as that can lead to undesirable memory
pressure and request-level latency.

This change creates a dedicated pair of configurations for the detection
timeout, decoupled from the dispatch timeout, and sets its default to
10s.