identity: add `identity_cert_refresh_count` metric (#788)
This commit adds a new `identity_cert_refresh_count` counter metric,
which tracks the number of times the proxy's identity certificate has
been successfully refreshed by the Identity service.

This change was fairly straightforward, but I did make a few internal
changes so that the counter could be passed into the daemon task. This
resulted in a couple things moving around, but the moved code is largely
unchanged.

In the future we may want to count cert refresh errors as well, possibly
adding labels for successes/failures, or a new metric tracking errors.