Simplify drain module implementation
The drain module can be updated to take advantage of the new `Notify`
primitive, replacing the more costly `watch` channel.

Before doing this, we update the module's tests to avoid manual future
polling. This change also updates method names for clarity,
elimintates an unnecessary manual future implementation, and eliminates
unnecessary crate dependencies.

The `drain::Retain` service is now gated behind a feature flag so the
tower and stack dependencies are optional. This is mostly helpful for
external crates that take a Git dependency on the drain crate.