Include scope names in failfast error messages (#791)
When a `FailFast` error is triggered, we have no way to know from which stack
this error originates.

This change modifies the `FailFast` module to accept a `scope` name to
be included in error log messages.

Furthermore, this change removes the `FailFastLayer` type in favor of a
simple `FailFast::layer` helper. It also removes the `push_failfast`
stack helper as it doesn't really help us (and the layer type is no
longer nameable).