inbound: Decouple inbound stack from TCP connections
In order to make the inbound stack testable without initializing OS
sockets, this change decouples the inbound stack from its TCP client,
passing in an abstract implementation to `inbound::Config::build`.

This change introduces a new `stack::Fail` module that always fails.
This is used to ensure that connections targeting the inbound server
without an opaque port are not forwarded, removing the need for
endpoint-level loop prevention.