Decouple TLS detection from TCP connections (#818)
The TLS accept stack--the `DetectTls` type--is currently coupled to the
`TcpStream` type; but the `TcpStream` type requires an actual OS-level
TCP connection, which isn't ideal for testing.

This change introduces a new trait `tls::accept::Detectable` and an
implementation for `TcpStream`. This will permit us to use alternate
implementations (e.g., for `io::DuplexStream`).

This change also updates the `DetectTls` type name to `NewDetectTls`,
and `AcceptTls` to `DetectTls`, to fit our more recent idioms.