replace `FutureService` with Tower's upstream impl (#839)
Tower 0.4 included a `FutureService` type which is basically identical
to the version we implemented for converting a `MakeService` into a
`NewService`. This branch replaces our implementation with the upstream
implementation from Tower.

It was necessary to add `MapErr` layers to convert the error types into
boxed errors, since our `FutureService` did that automatically, while
Tower's does not.