integration: improve tracing in tests (#846)
In stack tests, we set a fairly verbose `tracing` filter, since the logs
are captured by libtest and only printed if a test fails or when the
`--show-output` flag is passed. However, in integration tests, we
disable most tracing, because earlier versions of Rust did not support
output capturing for spawned threads, and the integration tests run the
proxy in a separate thread (see #771).

Now that we're on Rust 1.49, however, libtest now supports output
capturing for spawned threads, so this PR changes the integration tests
to enable more verbose logging. This should help with debugging test
failures.

I made a few other minor improvements to tracing in tests. In
particular:
* Share the *same* tracing subscriber between the test thread and the
  test proxy. This should resolve the issue where logs from the proxy
  in an integration test and logs from the test code have timestamps
  that are relative to different start times, making the logs appear
  to time travel.
* Replace some `println!`s with tracing events.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>