Annotate socket-level errors with a scope
We frequently encounter socket errors like:

    Connection closed error=Transport endpoint is not connected (os error 107)

While the handler is generally scoped with tracing context to tell us
information about the server-side connection, we lack information about
which socket actually failed. For every "logical" TCP stream, we
generally have two concrete connections: One on the server-side of the
proxy and another on the client-side.

This change introduces a new `io::ScopedIo` wrapper type that annotates
all I/O error messages with a `client` or `server` prefix. Then, the
proxy's clients and servers are wrapped with this type so that all
socket-level error messages are disambiguated.