tls: Disambiguate client and server identities
The `tls::PeerIdentity` type is used to describe both remote clients and
servers. This can easily lead to confusion, as it can be ambiguous as to
whether an identity is a client's identity or a target server's
identity.

This change introduces new marker types:

- `identity::LocalId`: The local proxy's ID;
- `tls::server::ClientId`: A remote client ID; and
- `tls::client::ServerId`: A target server ID.

Furthermore, the `tls::ReasonForNoPeerName` has been split into distinct
`tls::server::NoClientId` and `tls::client::NoServerId` types. This
change eliminates the `tls::HasPeerIdentity` and `tls::{client,
server}::HasConfig` types, in favor of simple `Into` coercions.

This change requires changes to the metric labeling.