Ensure services in failfast can become ready (#858)
When a Service is in failfast, the inner service is only polled as new
requests are processed. This means it's theoretically possible for
certain service tasks to be starved.

This change ensures that these layers are paired with a `SpawnReady`
layer to ensure that the inner service is always driven to readiness.

This could potentially explain behavior as described in
linkerd/linkerd2#5183; though we don't have strong evidence to support
that. This seems like a healthy defensive measure, in any case.

This change also improves stack commentary to favor larger descriptive
comments over layer-level annotations.

While auditing services for readiness, an unnecessary buffer has been
removed from the ingress HTTP stack.

Finally, this change updates the inbound connect timeout to 300ms to
accommodate especially slow endpoints.