test: reduce repetition in metrics tests
Currently, the metrics tests have a bunch of code that's repeated
between tests in the inbound and outbound directions. This makes
modifying the tests much more painful, since the code needs to be
changed multiple times. Also, the file is hard to navigate.

This commit changes this by factoring out the shared code into functions
which are called by the differing code (e.g. metrics labels) as
parameters. This means that the common behavior is only defined once, so
it should be somewhat easier to edit. This has a significant advantage
over the approach we use elsewhere, where similar tests are generated by
macros: it doesn't mess up the compiler diagnostics for errors in the
tests. We may want to refactor other tests to use a similar style in the
future.