inbound: Avoid HTTP detection when a transport header is present (#867)
In 1a7de34a, the inbound proxy was updated to hold a unified `direct`
stack; but in this change we started doing HTTP protocol detection even
when a transport header is present. But, currently, the existence of a
transport header indicates that the connection is likely opaque and so
such protocol detection must not be performed.

This change updates the direct stack to avoid HTTP detection when a
transport header is present.