admin: stackify admin service
This branch refactors the proxy admin endpoint to build a server stack
using the same primitives used for the proxy servers. This is intended
primarily to allow us to easily add metrics to the admin server by just
slotting in the existing proxy metrics layers.

This change also has some side benefits:

* We now reuse more existing code for building the admin server, rather
  than writing separate versions of our accept & serve HTTP code
  specifically for admin,
* We now support HTTP/2 for the admin server :)