tls: Report SNI value for non-Linkerd TLS
The proxy's TLS server parses the SNI value for all TLS connections.
This value is immediately discarded; but it's valuable information we
may want to repor to users.

This change modifies the TLS server to extract non-local SNI values and
pass them to the server as a stack parameter. This value is now exposed
as a metric & tap label, paired with a `tls="opaque"` label.

This change sets up extracing additional information on the TLS server
like ALPN values.