stack: Introduce the Param trait
When configuring a stack module, we are usually generic over the target
type, yet we need static assurances about the target's ability to
produce configuration values. Our current approach is to use awkward
type constraints like:

    for<'t> &'t T: Into<Config>

which is invoked as:

    (&target).into()

This is cumbersome and doesn't clearly express intent, especially in
implementations.

This change formalizes this pattern under a dedicated trait,
`stack::Param`. Stack modules are updated to express constraints like:

    T: Param<Config>

and is invoked as:

    target.param()

This change includes one minor change in functionality: the identity
client's TLS configuration is now obtained the module's `LocalCrtKey`
instead of its `TrustAnchors`. When using `TrustAnchors`, the client
would never initiate a client-authentcated mTLS session to the identity
service. When using the `LocalCrtKey`, the local identity may be used to
establish mTLS sessions when refreshing existing certificates.