transport-header: Encode session protocol
In order to transport both opaque and non-opaque protocols with the
transport header, especially for multicluster gateways, the inbound
proxy needs to be able to discover whether the session protocol is known
by the client.

This change adds a field and enum type to the transport header protobuf
message and wires it through so the outbound proxy parameterizes its
target type on the session protocol.a

Inbound handling of this field is not yet implemented and will be added
in a followup change.