stacK: Generalize switch to replace unwrap
When switching between stacks based on target values, we frequently want
to alter the target type so that the primary and secondary stacks may be
parameterized on different target types. This is the case with the
`unwrap_or` stack helper, for instance.

This change modifies the `Switch` trait to take the target by value and
return an `Either` indicating which stack should be used. Then, we
replace `NewUnwrapOr` with a `switch::Unwrap` helper.