test: keep profile resolutions alive in proxy-to-proxy tests
Currently, the proxy-to-proxy tests keep the Destination sender alive so
that the test controller doesn't close the destination resolution when
it's dropped, but they don't do this for profile resolutions. This means
logs from these tests are spammed with a ton of messages about proxies
trying to reconnect their profile resolutions.

This fixes that by holding the profile senders as well.