stack: Replace switch with Filter and NewEither (#880)
When switching between stacks based on target values, we frequently want
to alter the target type so that the primary and secondary stacks may be
parameterized on different target types. This is the case with the
`unwrap_or` stack helper, for instance.

This change removes the `Switch` trait and module, replacing uses with a
mix of `Fitler`, which modifies target types, and a new `NewEither` helper,
which takes `Either`-typed targets.

This change renames `NewUnwrapOr` as `UnwrapOr` and simplifies its
implementation to use `Filter` and `NewEither`

This change also removes various unused `MakeService` implementations.