inbound: normalize URIs after downgrading to HTTP/1 (#881)
Somewhere between stable-2.8.1 and stable-2.9.0, we introduced a bug
when downgrading HTTP/1 traffic that originated in origin form. See
linkerd/linkerd2#5298 for details.

Hyper's `Client` type requires all URIs to be in absolute form when
constructing the request, because the URI is used as a key for the
connection pool and to discover connect targets, but will write the
request out on the wire in origin-form unless explicitly configured not
to. Therefore, we normalize all HTTP/1 URIs to absolute-form before
handing them over to Hyper's `Client`, but only configure it to actually
send the request in absolute form if the request was received in
absolute-form. The `NormalizeUri` layer performs the absolute form
normalization and sets a request extension of the request was in
absolute-form originally.

The bug is that, when an orig-proto downgrade request is received in
origin form, it no longer gets normalized. It turns out that this is
because the `NormalizeUri` layer is in the inbound stack _above_ the
`orig_proto::Downgrade` layer --- so the request is still HTTP/2 when it
hits `NormalizeUri`, and is therefore not normalized.

This branch fixes the issue by separating the normalization from setting
the `WasAbsoluteForm` request extension, and moving the `NormalizeUri`
layer _after_ downgrade. The request extension is now set by a new
`DetectAbsoluteForm` layer.

I've also added new stack tests for inbound orig-proto requests in both
absolute- and origin-form. The origin-form test would have found the
issue, since it hits a debug assertion that requests are in
absolute-form, which none of the existing tests hit. Also, since inbound
didn't previously add any stack tests, I also added a simple unmeshed
http1 hello world test as well. This was mainly added as part of the
process for writing other stack tests: I wanted to have a working basic
request test before trying to write a test to reproduce the specific
issue.

Finally, I generalized a bunch of the utilities that were written for
the outbound stack tests so that they can also be used in inbound.

Fixes linkerd/linkerd2#5298.