metrics: Report build_info metric last (#882)
When spot-checking proxy metrics, it's extraordinarily helpful to be
able to spot-check the proxy's build info by having it be the last line
in the metrics output. #787 introduced identity metrics, appending them
to the end of the report.

This change moves identity metrics to the start of the metrics dump so
that build_info remains the final metric emitted.