Backport 2a645b72 to v2.124.0 (#884)
Between stable-2.8.1 and v2.124.0, the inbound proxy changed so
that orig-proto requests would not be properly normalized,
breaking compatibility between 2.8.1 clients and recent servers.

This was fixed in 2a645b72 on main, but main includes many other
changes since v2.124.0. This change backports the important parts
of the change to v2.124.0 to ensure that inbound proxies normalize
downgraded requests.