http: Parameterize the HTTP server
The HTTP server requires that targets are in the form `(Version, T)`.
This is inflexible, though.

This change updates the HTTP server to instead require that targets
satisfy `Param<Version>` so that tuple types are not required. This
requires the introduction of a new HttpAccept type in the inbound proxy.